requirements.txt
opencv-python>=4.9.0.80
numpy>=1.26.4
Pillow>=10.3.0
tqdm>=4.66.4
python-dotenv>=1.0.1
openai>=1.35.10
